TransactionReview: take state and functionality from TransactionReviewSummary, pass props
